adjustments after dev merge
